Move reaction handling from ListManager class to ListReactionHandler.
